backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
PyMuPDF>=1.24.0
google-generativeai>=0.8.0
fhir.resources>=7.1.0
python-dotenv>=1.0.0
pydantic>=2.10.0
aiofiles>=23.2.1
Pillow>=10.0.0
paddleocr>=2.8.1
numpy>=1.26.0
supabase>=2.0.0
fpdf2>=2.7.0
